refactor: ESC 키보드 이벤트 커스텀 훅으로 분리
