+++ components/src/lib/hooks/useEscapeKey.ts
@@ -1,0 +1,14 @@
+import { useEffect } from "react";
+
+const useEscapeKey = (onClose: () => void) => {
+  useEffect(() => {
+    const handleEscape = (e: KeyboardEvent) => {
+      if (e.key === "Escape") onClose();
+    };
+
+    document.addEventListener("keydown", handleEscape);
+    return () => document.removeEventListener("keydown", handleEscape);
+  }, [onClose]);
+};
+
+export default useEscapeKey;
